+++ db/pvault.sql
@@ -31,7 +31,7 @@ CREATE TABLE `bottles` (
   `supplier` varchar(255) COLLATE utf8_general_ci DEFAULT NULL,
   `supplier_link` varchar(255) COLLATE utf8_general_ci DEFAULT NULL,
   `notes` text COLLATE utf8_general_ci DEFAULT NULL,
-  `photo` varchar(255) COLLATE utf8_general_ci DEFAULT NULL
+  `pieces` int(11) NOT NULL DEFAULT 0
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 DROP TABLE IF EXISTS `customers`;